Refactor mobile application structure: implement authentication, event, organization, and Facebook routes; add controllers and models; remove unused PDF conversion logic.
