+++ node-app/middlewares/authMiddleware.js
@@ -1,10 +1,9 @@
 const jwt = require("jsonwebtoken");
 require("dotenv").config();
-const { Auth } = require("../models/userIdModel");
+const { Auth } = require("../mobile/models/userIdModel");
 
 const authMiddleware = (req, res, next) => {
-  const token = req.headers["authorization"]?.split(" ")[1]; // Extract token from "Bearer <token>"
-
+  const token = req.headers["authorization"]?.split(" ")[1];
   if (!token) {
     return res.status(401).json({ message: "No token provided" });
   }
